Replicator checkScalingThreshold now writes persistant state.
The introduction of ScalingThreshold meant we had additional
state fields. When these in-memory ints were getting updated, the
change was not being persisted to Consul. Therefore upon state
read the in-memory vaules were overwritten with the starting
values.

The checkScalingThreshold function will now finish by writing
update state to the persistant store, meaning evaluations are not
over-written or thrown away.

Closes #134